* master:
  use boost 1.67 (#24)
  Feature/merge to upstream (#23)

# Conflicts:
#	audiostream/test/ni/media/audio/aiff/aiff_sink.test.cpp
#	audiostream/test/ni/media/audio/wav/wav_sink.test.cpp

+++ audiostream/test/ni/media/audio/aiff/aiff_sink.test.cpp
@@ -45,7 +45,7 @@ TEST_P( aiff_sink_test, ofstream )
 
     {
         audio::ofstream os( output_name(), info );
-        auto            buffer = std::vector<int32_t>( is.info().num_samples() );
+        auto            buffer = std::vector<double>( is.info().num_samples() );
         is >> buffer;
         os << buffer;
     }
